improve array duplicates test
Remove duplicate assertion

In the remaining assertion you can clearly see that there are
three elements in the array.

In my mind, removing this assertion removes a bit of duplication
in the tests. This makes them a bit better than they were before.

Futher simplify remaining assertion

Removing the join() preserves the visibility of the return type
of the function.

[close #132]